+++ example/Nix/nixos-host/flake.nix
@@ -2,11 +2,11 @@
   description = "The configuration for nixos-host.";
 
   inputs = {
-    nixpkgs.url = "github:NixOS/nixpkgs/nixos-25.05";
-    confkit.url = "github:ejpcmac/confkit/stable-25.05";
+    nixpkgs.url = "github:NixOS/nixpkgs/nixos-25.11";
+    confkit.url = "github:ejpcmac/confkit/stable-25.11";
 
     home-manager = {
-      url = "github:nix-community/home-manager/release-25.05";
+      url = "github:nix-community/home-manager/release-25.11";
       inputs.nixpkgs.follows = "nixpkgs";
     };
   };