+++ lib/default.nix
@@ -2,5 +2,5 @@
   /**
     The current channel of confkit.
    */
-  channel = "develop-25.05";
+  channel = "stable-25.11";
 }